Fix incorrect BR packet throughput plots
The packet per second throughput in the border router evaluation was
incorrectly scaled by a factor of eight. The packet generator and raw
measurements are unaffected, but the diagrams pkts_1seg and pkt_2seg had
incorrect values.

The plotting script was erroneously multiplying both the number of bytes
(to obtain bits) and the number of packets per second with eight. This
patch corrects the scaling for the number of packets per second and
updates the diagrams. The new diagrams show the correct lower values,
but note that the relative performance of the router implementations is
not changed as they all have been affected by the same mistake. The
throughput in bytes per seconds is also not affected and shows correct
values in all plots.